refactor(turn-orchestrator): add failed state, work field; own entry types
